new make target to prepare a clientpkg locally
Install Perl libs that are needed as a clientpkg,
but compile all locally without chroots.

It's a poor man's clientpkg as it simply packs together
everything in /opt/tapper but can be used to build a
clientpkg by using Tapper-Deployment on a separate machine
with a wanted architecture and just running:

  make clientpkglocal

This creates a file 'opt-tapperlocal.tar.gz' in the same
path as the 64/32bit x86 clientpkgs, eg. in

  /data/tapper/live/repository/packages/tapperutils/

If you run it on your central Tapper host you get a lot
more from /opt/tapper but it doesn't hurt - just makes
the clientpkg bigger.

Essentially all this is just the Perl-related part of the
utility script 'prepare-clientpkg-in-chroot' and therefore
mostly useful to users of the public github version,
without the internal OSRC testsuites.