nfsroot: disable umount NFS annoyances
The NFSROOT is booted to run the installer
which sets up the machine and then reboots.

Sometimes the Debian config is broken and shuts
network early so NFS mounts hang and stop the reboot.

Now we drop the umount-nfs completely and don't
even stop networking at shutdown, so hopefully
the machine just shuts down without further ado.